feat: add ThemeProvider and TokenProvider for theme management
- Introduced a new `ThemeProvider` component that integrates with `next-themes` for managing light and dark themes.
- Implemented a `TokenProvider` to handle custom CSS variables and synchronize theme changes across documents.
- Added functionality for reading and persisting user-selected themes in local storage.
- Enhanced the context API to provide theme-related methods such as `setMode`, `mergeTheme`, and `reset` for better theme customization.
